requirements.txt
Flask==3.1.3
supabase==2.28.2
python-dotenv==1.2.2
httpx==0.28.1
gunicorn==23.0.0
